include doc string for consumer_go_router
